make all API routes restful and clean up documentation
