+++ controllers/api.js
@@ -4,29 +4,6 @@ var spotify = require('../config/spotifyApiHelper');
 var locus   = require('locus');
 var async   = require('async');
 
-var indexCircle = function(req, res) {
-  Circle.find({}, function(err, records) {
-    res.json(records);
-  });
-};
-
-var showCircle = function(req, res){
-  var id = req.params.id;
-  Circle.findById(id, function(err, circle){
-    console.log("GOT TO THIS");
-    if (err) {
-      res.send(err);
-    }
-    res.json(circle);
-  });
-};
-
-var indexUser = function(req, res) {
-  User.find({}, function(err, records) {
-    res.json(records);
-  });
-};
-
 var displayCircleUsers = function(req, res) {
     Circle.find({'_id': req.query._id }).populate('users').exec(function(err, circle){
         console.log(circle);
@@ -37,8 +14,5 @@ var displayCircleUsers = function(req, res) {
 
 
 module.exports = {
-  indexCircle: indexCircle,
-  showCircle: showCircle,
-  indexUser: indexUser,
   displayCircleUsers: displayCircleUsers
 }
