+++ controllers/users.js
@@ -1,13 +1,10 @@
 // Require resource's model(s).
 var User = require("../models/user");
 
-var index = function(req, res, next){
 
-  User.find({}, function(error, users){
-    res.render(
-      'users/index', {users: users}
-      // users
-    );
+var index = function(req, res) {
+  User.find({}, function(err, records) {
+    res.json(records);
   });
 };
 
@@ -18,7 +15,16 @@ var show = function(req, res, next){
   });
 };
 
+var currentUser = function(req, res, next){
+  User.findById(req.user.id, function(error, user){
+    if (error) res.json({message: 'Could not find user because ' + error});
+    res.send({user});
+  });
+}
+
+
 module.exports = {
   index: index,
-  show:  show
+  show:  show,
+  currentUser: currentUser
 };
