+++ views/index.ejs
@@ -2,6 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8">
+
     <$- include('partials/head') $>
   </head>
   <body>
